[clang] Move `AvailabilityInfo` into AST library
Previously this class was only used by ExtractAPI, but it will soon also
be needed by InstallAPI. This patch should not change availbility
behavior but just centralizes the information next to what already is
captured about availability for ast traversal.
